MC-33900: Issue with oversell of products
